+++ CHANGELOG.md
@@ -116,6 +116,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Email signature and layout (@mkasztelnik, @abacz)
 - Related Infrastructures and Platforms to filter section of services list (@martaswiatkowska)
 - Related Infrastructures and Platforms to service details view (@kmarszalek)
+- Technical parameters display in service offers (@michal-szostak)
 
 ### Changed
 - Upgrade Sprockets gem to avoid CVE-2018-3760 vulnerability (@mkasztelnik)